core/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.2
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
gunicorn==20.1.0
idna==3.4
mysql==0.0.3
mysqlclient==2.1.1
Pillow==9.2.0
psycopg2==2.9.4
PyJWT==2.5.0
python-decouple==3.6
pytz==2022.4
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
whitenoise==6.2.0
